requirements.txt
absl-py==2.3.1
accelerate==0.30.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
bert-score==0.3.13
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
datasets==2.19.0
dill==0.3.8
faiss-cpu==1.7.4
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.7.0
fsspec==2024.3.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.5
nltk==3.9.2
numpy==1.26.4
packaging==25.0
pandas==2.3.3
peft==0.10.0
pillow==11.3.0
propcache==0.3.2
psutil==7.1.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rouge_score==0.1.2
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.5.1
scipy==1.16.2
sentence-transformers==2.7.0
sentencepiece==0.2.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.8.0
tqdm==4.67.1
transformers==4.41.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
